chore: move net msg modules to its own dir
